+++ test-website/src/Pages/Test/RedirectingPage.tsx
@@ -1,15 +1,6 @@
 ﻿import React from 'react'
-import { Link } from 'react-router-dom'
-
 import { PageProps } from 'Components/Routing/RouteProps'
 import { NavbarLink } from 'Components'
-import {
-    SubmitButton,
-    Pager,
-    ActionDialog,
-    confirm,
-    ConfirmDialog
-} from '@interface-technologies/iti-react'
 
 interface PageState {
     shouldRedirect: boolean
@@ -17,7 +8,7 @@ interface PageState {
 
 /* This is to test that the page titles update correctly when a page pushes to history
  * in componentDidMount(), like a log out page would. */
-export class Page extends React.Component<PageProps, PageState> {
+export default class Page extends React.Component<PageProps, PageState> {
     state: PageState = {
         shouldRedirect: false
     }