+++ test-website/src/Pages/Test/TestRoutes.tsx
@@ -1,34 +1,28 @@
-﻿import {
+﻿import React from 'react'
+import {
     RoutesProps,
     LocalRoutesProps,
     passPageProps
 } from 'Components/Routing/RouteProps'
 import { getProtectedRouteBuilder } from 'Components/Routing/ProtectedRoute'
-import { CustomLoadable } from '@interface-technologies/iti-react'
 
-const Form = CustomLoadable(() => import('./Form').then(m => m.Page) as any) as any
-const Components = CustomLoadable(
-    () => import('./Components/Components').then(m => m.Page) as any
-) as any
-const Inputs = CustomLoadable(
-    () => import('./Inputs/Inputs').then(m => m.Page) as any
-) as any
-const RouteParam = CustomLoadable(
-    () => import('./RouteParam').then(m => m.Page) as any
-) as any
-const RedirectingPage = CustomLoadable(
-    () => import('./RedirectingPage').then(m => m.Page) as any
-) as any
-const TabManager = CustomLoadable(
-    () => import('./TabManager/TabManager').then(m => m.Page) as any
-) as any
-const UrlSearchParam = CustomLoadable(
-    () => import('./UrlSearchParam').then(m => m.Page) as any
-) as any
-const Hooks = CustomLoadable(() => import('./Hooks').then(m => m.Page) as any) as any
-const SpamOnReady = CustomLoadable(() =>
-    import('./SpamOnReady').then(m => m.Page as any)
-) as any
+const Form = React.lazy(() => import('./Form'))
+const Components = React.lazy(() => import('./Components/Components'))
+const Inputs = React.lazy(() => import('./Inputs/Inputs'))
+const RouteParam = React.lazy(() => import('./RouteParam'))
+const RedirectingPage = React.lazy(
+    () => import('./RedirectingPage')
+)
+const TabManager = React.lazy(
+    () => import('./TabManager/TabManager')
+)
+const UrlSearchParam = React.lazy(
+    () => import('./UrlSearchParam')
+)
+const Hooks = React.lazy(() => import('./Hooks'))
+const SpamOnReady = React.lazy(() =>
+    import('./SpamOnReady')
+)
 
 export const paths = {
     routeParam: '/test/routeParam/:number'